Refactor to support calling via a clojure -X :exec-fn.  Closes #16
We rename the deploy multi-method to deploy* and consequently there is
a chance this may break any users who were extending this multimethod,
so we can wrap this with a new deploy function.

Also renames :clojars key to :remote to indicate the mechanism for
repositories is more generic than just clojars.